fetch: 병합중
