+++ src/app/scenario/page.js
@@ -5,7 +5,7 @@ import Box from "@mui/material/Box";
 
 export default function ScenarioPage() {
   return (
-    <Box sx={{ p: 2 }}>
+    <Box>
       <DataGridSection />
     </Box>
   );
